model/requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
evaluate>=0.4.0
scikit-learn>=1.3.0
onnx>=1.15.0
onnxruntime>=1.16.0
optimum>=1.16.0
captum>=0.7.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0
matplotlib>=3.8.0
seaborn>=0.13.0
